Adding a way to get functions from the custom property on the tmx
At first i tried implemeting it with _FUNCTION_ macro but it only returns the name of the function that is executing. And it isn't a good idea to execute the ExitGame function just to know its name.
The solution I'm implementing atm is based on a std::map which associates a string value with a function pointer value.
When a button is created, it passes the string from the tmx and calls a function called GetButtonFunction which returns the pointer given the name.