Commented gamepad support
It seems that the gamepad support we added was causing the crashes in libraries we didn't have the source code, so it was extremely hard to debug. In this particular case er can't explain how we found the fix more than that we vaguely remember that it started crashing around the time we incorporated the gamepad support and some other things.
We commented it for now and we'll test what is the specific problem that's causing the crashes.